materials can be only downloaded if signed in
